# docs/openapi.yaml
openapi: 3.0.3
info:
  title: パートナー協業プラットフォーム API
  description: |
    パートナー企業との協業を管理するためのREST API仕様です。

    ## 認証
    このAPIはBearer Token (JWT)認証を使用します。
    ログインAPIで取得したアクセストークンを`Authorization`ヘッダーに設定してください。

    ```
    Authorization: Bearer {accessToken}
    ```

    ## ページネーション
    一覧取得APIでは`limit`と`offset`パラメータでページネーションを制御できます。

    ## エラーレスポンス
    全てのエラーは統一フォーマットで返却されます。
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
servers:
  - url: https://api.example.com/api
    description: Production server
  - url: https://staging-api.example.com/api
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /reminders:
    get:
      tags:
        - Reminders
      summary: リマインド設定一覧取得
      description: リマインド設定一覧を取得します。
      operationId: listReminders
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: isActive
          in: query
          description: 有効なリマインドのみ取得
          schema:
            type: boolean
        - name: type
          in: query
          description: タイプでフィルタ
          schema:
            $ref: '#/components/schemas/ReminderType'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Reminders
      summary: リマインド設定作成
      description: 新規リマインド設定を作成します。
      operationId: createReminder
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderCreateRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Reminder'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWTアクセストークン
  parameters:
    LimitParam:
      name: limit
      in: query
      description: 取得件数
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    OffsetParam:
      name: offset
      in: query
      description: 取得開始位置
      schema:
        type: integer
        minimum: 0
        default: 0
...
  responses:
    Unauthorized:
      description: 認証エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: 認証が必要です
...
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: 入力内容に誤りがあります
              details:
                - field: email
                  message: 有効なメールアドレスを入力してください
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: エラーコード
            message:
              type: string
              description: エラーメッセージ
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: フィールド名
                  message:
                    type: string
                    description: 詳細メッセージ
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: 総件数
        limit:
          type: integer
          description: 取得件数
        offset:
          type: integer
          description: 取得開始位置
        hasMore:
          type: boolean
          description: 追加データの有無
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          description: ユーザーID
        email:
          type: string
          format: email
          description: メールアドレス
        name:
          type: string
          description: 氏名
        role:
          type: string
          enum: [admin, manager, member, partner]
          description: 役割

    # ============================================================
    # Project Schemas
    # ============================================================
...
    ReminderType:
      type: string
      enum: [task, project, custom]
      description: リマインドタイプ
    Reminder:
      type: object
      properties:
        id:
          type: string
          description: リマインドID
        type:
          $ref: '#/components/schemas/ReminderType'
        title:
          type: string
          description: タイトル
        description:
          type: string
          description: 説明
        isActive:
          type: boolean
          description: 有効フラグ
        triggerCondition:
          $ref: '#/components/schemas/TriggerCondition'
        notificationChannels:
          type: array
          items:
            type: string
            enum: [email, in_app, slack]
          description: 通知チャネル
        targetUsers:
          type: array
          items:
            type: string
            enum: [assignee, manager, all_members]
          description: 通知対象
        relatedProjectId:
          type: string
          nullable: true
          description: 関連案件ID
        relatedTaskId:
          type: string
          nullable: true
          description: 関連タスクID
        schedule:
          $ref: '#/components/schemas/Schedule'
        lastTriggeredAt:
          type: string
          format: date-time
          nullable: true
          description: 最終実行日時
        nextTriggerAt:
          type: string
          format: date-time
          nullable: true
          description: 次回実行日時
        createdBy:
          $ref: '#/components/schemas/UserSummary'
        createdAt:
          type: string
          format: date-time
          description: 作成日時
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
    ReminderListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Reminder'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ReminderCreateRequest:
      type: object
      required:
        - type
        - title
        - notificationChannels
        - targetUsers
      properties:
        type:
          $ref: '#/components/schemas/ReminderType'
        title:
          type: string
          maxLength: 100
          description: タイトル
        description:
          type: string
          maxLength: 500
          description: 説明
        isActive:
          type: boolean
          default: true
          description: 有効フラグ
        triggerCondition:
          $ref: '#/components/schemas/TriggerCondition'
        notificationChannels:
          type: array
          items:
            type: string
            enum: [email, in_app, slack]
          description: 通知チャネル
        targetUsers:
          type: array
          items:
            type: string
            enum: [assignee, manager, all_members]
          description: 通知対象
        relatedProjectId:
          type: string
          description: 関連案件ID
        relatedTaskId:
          type: string
          description: 関連タスクID
        schedule:
          $ref: '#/components/schemas/Schedule'
...
    TriggerCondition:
      type: object
      nullable: true
      properties:
        event:
          type: string
          enum: [task_due, project_end, milestone_due]
          description: イベントタイプ
        offsetDays:
          type: integer
          description: オフセット日数
        offsetHours:
          type: integer
          description: オフセット時間
        filter:
          type: object
          additionalProperties: true
          description: フィルタ条件
    Schedule:
      type: object
      nullable: true
      properties:
        type:
          type: string
          enum: [daily, weekly, monthly]
          description: スケジュールタイプ
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
          description: 曜日（0-6、週次の場合）
        dayOfMonth:
          type: integer
          minimum: 1
          maximum: 31
          description: 日（1-31、月次の場合）
        time:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$'
          description: 時刻（HH:mm形式）